+++ app/branding/page.tsx
@@ -1,9 +1,0 @@
-import { requireAdminAccess } from '../../src/lib/roleGate';
-import BrandingPage from './PageClient';
-
-export const metadata = { title: 'Branding & Deliverability' };
-
-export default async function Page() {
-    await requireAdminAccess();
-    return <BrandingPage />;
-}